Cleaning the code, adding some new endpoints

+++ src/main/scala/com/applaudo/akkalms/models/responses/IncomeResponse.scala
@@ -1,6 +1,6 @@
 package com.applaudo.akkalms.models.responses
 
-case class IncomeResponse(id: Int,
+case class IncomeResponse(id: Long,
                           incomeType: String,
                           amount: BigDecimal,
                           currency: String,
